Revert "Revert "Revert "Viewer (previously Monitor) role"""
