+++ dev/com.ibm.ws.jmx.connector.server.rest/resources/WEB-INF/web.xml
@@ -26,7 +26,7 @@
   </servlet-mapping>
 
   <security-constraint id="SecurityConstraint_1">
-    <display-name>REST Connector Security Constraint - Administrator Role</display-name>
+    <display-name>REST Connector Security Constraint</display-name>
     <web-resource-collection id="WebResourceCollection_1">
         <web-resource-name>REST Connector URL</web-resource-name>
         <url-pattern>/*</url-pattern>     
@@ -40,7 +40,7 @@
   </security-constraint>
 
   <security-constraint id="SecurityConstraint_2">
-    <display-name>REST Connector Security Constraint - OPTIONS</display-name>
+    <display-name>REST Connector Security Constraint</display-name>
     <web-resource-collection id="WebResourceCollection_2">
         <web-resource-name>REST Connector URL</web-resource-name>
         <url-pattern>/*</url-pattern>
@@ -51,29 +51,9 @@
     </user-data-constraint>
   </security-constraint>
 
-  <security-constraint id="SecurityConstraint_3">
-    <display-name>REST Connector Security Constraint - Viewer Role</display-name>
-    <web-resource-collection id="WebResourceCollection_3">
-        <web-resource-name>REST Connector URL</web-resource-name>
-        <url-pattern>/*</url-pattern>
-        <http-method>GET</http-method>
-    </web-resource-collection>
-    <auth-constraint>
-      <role-name>Viewer</role-name>
-    </auth-constraint>
-    <user-data-constraint id="UserDataConstraint_3">
-      <transport-guarantee>CONFIDENTIAL</transport-guarantee>
-    </user-data-constraint>
-  </security-constraint>
-
   <security-role id="SecurityRole_1">
     <description>Administrator Role</description>
     <role-name>Administrator</role-name>
-  </security-role>
-
-  <security-role id="SecurityRole_2">
-    <description>Viewer Role</description>
-    <role-name>Viewer</role-name>
   </security-role>
   
   <login-config id="LoginConfig_1">
