+++ dev/com.ibm.ws.management.security/resources/OSGI-INF/l10n/metatype.properties
@@ -14,12 +14,9 @@
 #NLS_ENCODING=UNICODE
 #NLS_MESSAGEFORMAT_NONE
 
-#==== MANAGEMENT ROLE CONFIGURATION =====
+#==== ADMINISTRATOR ROLE CONFIGURATION =====
 role.administrator=Administrator Role
-role.administrator.desc=A collection of users, groups, or both that are assigned the server administrator role.
-
-role.viewer=Viewer Role
-role.viewer.desc=A collection of users, groups, or both that are assigned the server viewer role.
+role.administrator.desc=A collection of users and/or groups assigned the server administrator role.
 
 role.user=User name
 role.user.desc=User assigned a role.
