+++ dev/com.ibm.ws.management.security/resources/OSGI-INF/metatype/metatype.xml
@@ -30,21 +30,4 @@
         <Object ocdref="com.ibm.ws.management.security.role.administrator.metatype" />
     </Designate>
 
-    <OCD id="com.ibm.ws.management.security.role.viewer.metatype" name="%role.viewer" description="%role.viewer.desc"
-         ibm:alias="viewer-role" ibm:objectClass="com.ibm.ws.management.security.ManagementRole">
-        <AD id="user" name="%role.user" description="%role.user.desc"
-            required="false" type="String" cardinality="2147483647" />
-        <AD id="user-access-id" name="%role.userAccessId" description="%role.userAccessId.desc"
-            required="false" type="String" cardinality="2147483647" />
-
-        <AD id="group" name="%role.group" description="%role.group.desc"
-            required="false" type="String" cardinality="2147483647" />
-        <AD id="group-access-id" name="%role.groupAccessId" description="%role.groupAccessId.desc"
-            required="false" type="String" cardinality="2147483647" />
-    </OCD>
-
-    <Designate pid="com.ibm.ws.management.security.role.viewer">
-        <Object ocdref="com.ibm.ws.management.security.role.viewer.metatype" />
-    </Designate>
-
 </metatype:MetaData>
