+++ dev/com.ibm.ws.management.security/resources/com/ibm/ws/management/security/internal/resources/LoggingMessages.nlsprops
@@ -25,7 +25,3 @@ MANAGEMENT_SECURITY_AUTHZ_FAILED.useraction=No action is required.
 ROLE_ENTRY_DUPLICATE=CWWKX0002E: A duplicate entry in the {0} role binding was detected. The duplicate {1} entry is {2}.
 ROLE_ENTRY_DUPLICATE.explanation=The specified entry is already defined for the specified role binding. Duplicate entries are not supported.
 ROLE_ENTRY_DUPLICATE.useraction=Remove the duplicate entry.
-
-INVALID_ACCESS_ID=CWWKX0003E: An invalid access ID in the {0} role binding was detected. The invalid {1} entry is {2}.
-INVALID_ACCESS_ID.explanation=The specified entry is not a valid access ID for the specified role binding.
-INVALID_ACCESS_ID.useraction=Remove or correct the invalid entry.
