+++ dev/com.ibm.ws.rest.handler/src/com/ibm/wsspi/rest/handler/helper/DefaultAuthorizationHelper.java
@@ -20,29 +20,20 @@ import com.ibm.wsspi.rest.handler.RESTResponse;
 
 /**
  * This helper service performs the default authorization on the given user.
- *
+ * 
  * @ibm-spi
  */
 @Component(service = { DefaultAuthorizationHelper.class }, configurationPolicy = ConfigurationPolicy.IGNORE, immediate = true, property = { "service.vendor=IBM" })
 public class DefaultAuthorizationHelper {
 
     public boolean checkAdministratorRole(RESTRequest request, RESTResponse response) throws IOException {
-        boolean isGetMethod = "GET".equals(request.getMethod());
-
         if (request.isUserInRole("Administrator")) {
             return true;
         }
-        if (isGetMethod && request.isUserInRole("Viewer")) {
-            return true;
-        }
 
-        // Not in admin role, so built error msg
-        // TODO: Translate msg
-        if (isGetMethod) {
-            response.sendError(403, "Administrator or Viewer role needed.");
-        } else {
-            response.sendError(403, "Administrator role needed.");
-        }
+        //Not in admin role, so built error msg
+        //TODO: Translate msg
+        response.sendError(403, "Administrator role needed.");
 
         return false;
     }
